Refactoring to remove somes duplicated code
